added new form to client component
